fix: コードレビュー指摘修正 + 新コマンド追加

+++ MapChooserSharpMS/Modules/Nomination/Commands/NomListCommand.cs
@@ -75,7 +75,7 @@ internal sealed class NomListCommand(IServiceProvider provider) : TnmsAbstractCo
 
         foreach (int slot in nomination.NominationParticipants)
         {
-            var participant = clientManager.GetGameClient(new PlayerSlot(slot));
+            var participant = clientManager.GetGameClient(new PlayerSlot((byte)slot));
             if (participant is null) continue;
 
             if (sb.Length > 0)
